requirements.txt
# Web & API
fastapi
uvicorn

# DB
psycopg
psycopg_pool
secrets
redis
sqlite3

# Data & Utilities
python-dotenv
numpy

# LLM
Ollama

# RAG / NLP
sentence-transformers
torch
transformers
scikit-learn

# Logging
logging
contextlib

# Optional (for typing / dev)
pydantic

